requirements.txt
Django==1.6.2
PIL==1.1.7
facebook-sdk==0.4.0
oauthlib==0.6.1
python-openid==2.2.5
python-social-auth==0.1.21
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
wsgiref==0.1.2
